# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Pool API documentation",
        "contact": {},
        "version": "3.0.0"
    },
    "host": "pool3-api.vivaconagua.org",
    "basePath": "/v1",
    "paths": {
...
        "/events/artist": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/events/artist"
                ],
                "summary": "Get a List of Artists",
                "parameters": [
                    {
                        "type": "string",
                        "name": "createdFrom",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "createdTo",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "updatedFrom",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "updatedTo",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/vcago.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "payload": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Artist"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/vcago.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/events/artist"
                ],
                "summary": "Get a Artist by ID",
                "parameters": [
                    {
                        "description": "Artist Data",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ArtistUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/vcago.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "payload": {
                                            "$ref": "#/definitions/models.Artist"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "creates an  Artist object.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/events/artist"
                ],
                "summary": "Create a Artist",
                "parameters": [
                    {
                        "description": "Artist Data",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ArtistCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/vcago.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "payload": {
                                            "$ref": "#/definitions/models.Artist"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/events/artist/{id}": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/events/artist"
                ],
                "summary": "Get a  Artist by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Artist ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/vcago.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "payload": {
                                            "$ref": "#/definitions/models.Artist"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/events/artist"
                ],
                "summary": "Get a  Artist by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Artist ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/vmod.DeletedResponse"
                        }
                    }
                }
            }
        },
...
        "/users/avatar": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "creates an  Avatar object.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/users/avatar"
                ],
                "summary": "Uploads a Avatar",
                "parameters": [
                    {
                        "description": "Avatar File",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AvatarFile"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/vcago.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "payload": {
                                            "$ref": "#/definitions/models.Avatar"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/users/avatar/{id}": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "image/png"
                ],
                "tags": [
                    "/users/avatar"
                ],
                "summary": "Get a  Avatar by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Avatar ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Download Image",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/users/avatar"
                ],
                "summary": "Get a  Avatar by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Avatar ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/vmod.DeletedResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Artist": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "modified": {
                    "$ref": "#/definitions/vmod.Modified"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.ArtistCreate": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "models.ArtistUpdate": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Avatar": {
            "type": "object",
            "properties": {
                "file_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "modified": {
                    "$ref": "#/definitions/vmod.Modified"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.AvatarFile": {
            "type": "object",
            "properties": {
                "file": {},
                "header": {
                    "$ref": "#/definitions/multipart.FileHeader"
                }
            }
        },
...
        "multipart.FileHeader": {
            "type": "object",
            "properties": {
                "filename": {
                    "type": "string"
                },
                "header": {
                    "$ref": "#/definitions/textproto.MIMEHeader"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
        "textproto.MIMEHeader": {
            "type": "object",
            "additionalProperties": {
                "type": "array",
                "items": {
                    "type": "string"
                }
            }
        },
...
        "vcago.Response": {
            "type": "object",
            "properties": {
                "list_size": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "payload": {},
                "type": {
                    "type": "string"
                }
            }
        },
...
        "vmod.DeletedResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "vmod.Modified": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "integer"
                },
                "updated": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Bearer key",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "CookieAuth": {
            "type": "apiKey",
            "name": "access_cookie",
            "in": "cookie"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://hydra.vivaconagua.org/oauth/authorize",
            "tokenUrl": "https://hydra.vivaconagua.org/oauth/token"
        }
    }
}